restructure live-chat
